+++ py/desitarget/io.py
@@ -22,7 +22,7 @@ tscolumns = [
     'DECAM_FRACFLUX', 'DECAM_FLUX_IVAR',
     'WISE_FLUX', 'WISE_MW_TRANSMISSION',
     'WISE_FLUX_IVAR',
-    'SHAPEDEV_R', 'SHAPEEXP_R',
+    'SHAPEDEV_R', 'SHAPEEXP_R','DCHISQ',
     ]
 
 def read_tractor(filename, header=False, columns=None):
@@ -60,10 +60,11 @@ def read_tractor(filename, header=False, columns=None):
     data = fx[1].read(columns=readcolumns)
 
     #ADM To circumvent whitespace bugs on I/O from fitsio
-    #ADM need to enforce that the TYPE column is 4 characters
-    #ADM only other string column is BRICKNAME, but its length is consistent
-    if 'TYPE' in readcolumns:
-        data["TYPE"] = np.char.ljust(data["TYPE"],4)
+    #ADM need to strip any white space from string columns
+    for colname in data.dtype.names:
+        kind = data[colname].dtype.kind
+        if kind == 'U' or kind == 'S':
+            data[colname] = np.char.rstrip(data[colname])
 
     if header:
         hdr = fx[1].read_header()
@@ -264,3 +265,34 @@ def check_fitsio_version(version='0.9.8'):
         not fitsio.__version__.startswith(version)):
         raise ImportError(('ERROR: fitsio >{0}rc1 required ' +
                            '(not {1})!').format(version, fitsio.__version__))
+
+def whitespace_fits_read(filename, **kwargs):
+    """Use fitsio_ to read in a file and strip whitespace from all string columns
+
+    .. _fitsio: https://pypi.python.org/pypi/fitsio
+
+    Parameters
+    ----------
+    filename : :class:`str`
+        Name of the file to be read in by fitsio
+    kwargs: arguments that will be passed directly to fitsio
+    """
+    fitout = fitsio.read(filename, **kwargs)
+    #ADM if the header=True option was passed then
+    #ADM the output is the header and the data
+    data = fitout
+    if len(fitout) == 2:
+        data, header = fitout
+
+    #ADM guard against the zero-th extension being read by fitsio
+    if data is not None:
+        #ADM strip any whitespace from string columns
+        for colname in data.dtype.names:
+            kind = data[colname].dtype.kind
+            if kind == 'U' or kind == 'S':
+                data[colname] = np.char.rstrip(data[colname])
+
+    if len(fitout) == 2:
+        return data, header
+
+    return data